src/summary_vocab/requirements.txt
google-cloud-aiplatform
google-cloud-storage
pytest
pytest-cov
